refactor: refactor router code and improve frontend page
